Initial work on porting the Ruby version to Python.
So far the parser class has been ported. I need to work out
how I want to implement an equivalent to the Ruby Generator
class as it's doing dynamic code generation which I've not done
too much of. Should be too hard, I just need to figure out if
that's the direction I want to go.