Mostly implemented refactoring.
I've got enough written that I can use views with string templates. Reading
through sources to Ruby and the old Python code, I'm not a big fan of the way
that class variables are used to find and load template code form the file
system. Mako's TemplateLookup is nice, but feels a bit more complicated than
need be. There's also the case where I may want to use a single View with
multiple templates. I feel as though there's a simple way to accomplish this,
but I haven't yet thought of it.