Implement the {{^}} syntax for else-sections.
This is a simple modification to the parser that allows a {{^}}
shorthand syntax. The Mustache 1.0 pattern would look like this:

    {{#section}}
    ... stuff if truthy ...
    {{/section}}
    {{^section}}
    ... stuff if falsy ...
    {{/section}}

This can now be abbreviated with the similar yet more concise syntax:

    {{#section}}
    ... stuff if truthy ...
    {{^}}
    ... stuff if falsy ...
    {{/section}}

Internally this is exactly the same as before. When parsing {{^}}
we just close the previous section and open a new inverted section.
Most importantly, the section variable will be evaluated twice just
as it was before.