Address enclosing contexts using {{foo^}} syntax.
Given a context and template like the following, we can now replace
the _HERE_ string with {{baz^}} or {{bar^^.baz}} to access either
of the two versions of baz.

    ctx = {
        "foo": {"bar": "bing", "baz": "bam"},
        "bar": {"baz": "bang"}
    }

    {{#foo}}{{#bar}}_HERE_{{/bar}}{{/foo}}

You can only use the ^ syntax on the first name in a dotted name
lookup. If you specify more ^'s than there are elements in the
current context stack, the root context is used.